requirements.txt
Flask
python-docx
reportlab
PyPDF2
fpdf